cloud/GCP/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cloudbooter-gcp"
version = "0.1.0"
description = "GCP Always Free Tier provisioning toolkit — parity with ./cloud/OCI"
requires-python = ">=3.11"
dependencies = [
  "google-auth>=2.38.0",
  "google-auth-httplib2>=0.2.0",
  "google-cloud-compute>=1.22.0",
  "google-cloud-storage>=2.19.0",
  "google-cloud-resource-manager>=1.14.0",
  "google-cloud-billing>=1.14.0",
  "google-cloud-secret-manager>=2.22.0",
  "google-cloud-logging>=3.11.0",
  "google-api-python-client>=2.164.0",
  "requests>=2.32.0",
  "click>=8.1.0",
  "jinja2>=3.1.0",
  "pyyaml>=6.0",
  "rich>=13.0.0",
  "python-terraform>=0.10.0",
]

[project.scripts]
cloudbooter-gcp = "cloudbooter.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.ruff]
line-length = 100
